removed old tests
