--- scrapedoutput/artists/dbpedia_test.json
@@ -1,1 +1,0 @@
-[{"name":"Frank Zappa","nationality":"American","dateOfBirth":"1940-12-21","dateOfDeath":"1993-12-04","placeOfBirth":"Baltimore, Maryland, U.S.","placeOfDeath":"Los Angeles, California, U.S.","instrument":null,"psuedonym":null,"work":null,"release":null,"tag":["American classical musicians","Contemporary classical composers","Contemporary classical music performers","Humor in classical music"],"source_link":"http://dbpedia.org/resource/Frank_Zappa","wikipedia_link":"http://en.wikipedia.org/wiki/Frank_Zappa","wikipedia_pageid":"10672"}]
