+++ scrapedoutput/artists/testartist.json
@@ -1,0 +1,1 @@
+[{"name":"David Shifrin","nationality":"American","dateOfBirth":null,"dateOfDeath":null,"placeOfBirth":null,"placeOfDeath":null,"instrument":null,"psuedonym":null,"work":null,"release":null,"tag":["American classical musicians"],"source_link":"http://dbpedia.org/resource/David_Shifrin","wikipedia_link":"http://en.wikipedia.org/wiki/David_Shifrin","wikipedia_pageid":"2798287"}]
