+++ scrapedoutput/releases/testrelease.json
@@ -1,0 +1,1 @@
+[{"title":"Release 1","format":"Vinyl","date":"1993-12-04", "country":"Germany","label":"Universal"},{"title":"Great Release 2","format":"CD","date":"2000-01-04", "country":"France","label":"Sony"}]
