+++ scrapedoutput/works/testwork.json
@@ -1,0 +1,1 @@
+[{"title":"Work 1","compositionyear":"1759"},{"title":"Great Work","compositionyear":"1990"}]
