--- testscripts/test.js
@@ -1,5 +1,0 @@
-/**
- * Created by TimH on 28-Mar-17.
- */
-console.log("test 1");
-process.send("done 1");
