--- testscripts/test2.js
@@ -1,6 +1,0 @@
-/**
- * Created by TimH on 28-Mar-17.
- */
-
-console.log("test 2");
-process.send("done 2");
