--- testscripts/test3.js
@@ -1,5 +1,0 @@
-/**
- * Created by TimH on 28-Mar-17.
- */
-console.log("test 3");
-process.send("done 3");
